requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
CacheControl==0.12.10
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.11
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
dj-rest-auth==2.2.3
Django==4.0.2
django-anymail==8.5
django-ckeditor==6.2.0
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-extensions==3.1.5
django-fcm==0.1.1
django-filter==21.1
django-hosts==5.1
django-import-export==2.7.1
django-js-asset==2.0.0
django-mdeditor==0.1.20
django-modeltranslation==0.18.2
django-redis==5.2.0
django-redis-sessions==0.6.2
django-tinymce==3.4.0
django-versatileimagefield==2.2
djangorestframework==3.13.1
djangorestframework-api-key==2.1.0
drf-yasg==1.20.0
et-xmlfile==1.1.0
fcm-django==1.0.8
firebase-admin==5.2.0
google-api-core==2.5.0
google-api-python-client==2.36.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-cloud-core==2.2.2
google-cloud-firestore==2.3.4
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
gunicorn==20.1.0
httplib2==0.20.4
huey==2.4.3
idna==3.3
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.0.3
MarkupPy==1.14
MarkupSafe==2.0.1
msgpack==1.0.3
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
Pillow==9.0.1
proto-plus==1.19.9
protobuf==3.19.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dateutil==2.8.2
python-magic==0.4.22
python-magic-bin==0.4.14
pytz==2021.3
PyYAML==6.0
redis==4.1.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tablib==3.2.0
tweepy==4.5.0
uritemplate==4.1.1
urllib3==1.26.8
wrapt==1.13.3
xlrd==2.0.1
xlwt==1.3.0
